pyproject.toml
[project]
name = "construction-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
